fix vibrating animation
Animation via css transform breaks flash content, so I removed transform while the slider was not animating and replaced the positioning with margin-left. But that causes a vibration while animating with transform again. The Video Players (YouTube/Vimeo) has to load dynamically and there HTML has to placed outside the slate of the slider to get this working.